Use iterative approach to test application of CompanyAuth middleware
