+++ src/controllers/ReviewPenaltyController.ts
@@ -25,8 +25,8 @@ const navigation = {
     }
 };
 
-@controller(REVIEW_PENALTY_PAGE_URI, SessionMiddleware, AuthMiddleware, CheckForDuplicateMiddleware,
-    CompanyAuthMiddleware)
+@controller(REVIEW_PENALTY_PAGE_URI, SessionMiddleware, AuthMiddleware, CompanyAuthMiddleware,
+    CheckForDuplicateMiddleware)
 export class ReviewPenaltyController extends SafeNavigationBaseController<PenaltyDetailsTable> {
 
     public static PENALTY_EXPECTED_ERROR: string = 'Penalty object expected but none found';
