+++ test/controllers/SelectPenaltyController.test.ts
@@ -1,6 +1,6 @@
 import { Penalty, PenaltyList } from 'ch-sdk-node/dist/services/lfp';
 import { expect } from 'chai';
-import { INTERNAL_SERVER_ERROR, MOVED_TEMPORARILY } from 'http-status-codes';
+import { INTERNAL_SERVER_ERROR } from 'http-status-codes';
 import request from 'supertest';
 import { createApp } from '../ApplicationFactory';
 
@@ -51,40 +51,6 @@ describe('SelectPenaltyController', () => {
                 expect(res.text).to.include('value="A0000002"');
             });
     });
-
-    it('GET: should redirect to Company Authentication when user is not authenticated', async () => {
-        const applicationData: Partial<ApplicationData> = {
-            appeal: {
-                penaltyIdentifier: {
-                    companyNumber: 'NI999999',
-                    penaltyReference: 'PEN1A/NI000000',
-                    companyName: 'Test',
-                    penaltyList: {
-                        items: [
-                            {} as Penalty,
-                            {} as Penalty
-                        ]
-                    } as PenaltyList
-                }
-            } as Appeal,
-            navigation: { permissions: [SELECT_THE_PENALTY_PAGE_URI] }
-        };
-
-        /* TODO: Work out how to use this function call to override default Company Number
-        This will make the test data more clear, instead of changing the CN in the applicationData
-        above */
-        const app = createApp(applicationData);
-
-        await request(app)
-            .get(SELECT_THE_PENALTY_PAGE_URI)
-            .expect(res => {
-                expect(res.status).to.be.equal(MOVED_TEMPORARILY);
-                expect(res.text).to.contain('Found. Redirecting')
-                    .and.to.contain('oauth2/authorise')
-                    .and.to.contain('scope=https://api.companieshouse.gov.uk/company/NI999999');
-            });
-    });
-
 
     it('POST: should show an error if the penalty list is undefined', async () => {
         const applicationData: Partial<ApplicationData> = {
