+++ test/middleware/CompanyAuthMiddleware.test.ts
@@ -4,8 +4,12 @@ import { Arg, Substitute} from '@fluffy-spoon/substitute';
 import { Session, SessionStore } from 'ch-node-session-handler';
 import { SessionKey } from 'ch-node-session-handler/lib/session/keys/SessionKey';
 import { ISignInInfo } from 'ch-node-session-handler/lib/session/model/SessionInterfaces';
+import { Penalty, PenaltyList } from 'ch-sdk-node/dist/services/lfp';
 import { expect } from 'chai';
 import { NextFunction, Request, Response } from 'express';
+import { MOVED_TEMPORARILY } from 'http-status-codes';
+import request from 'supertest';
+import { createApp } from '../ApplicationFactory';
 
 import 'app/controllers/index';
 import { CompanyAuthMiddleware } from 'app/middleware/CompanyAuthMiddleware';
@@ -14,6 +18,7 @@ import { ApplicationData, APPLICATION_DATA_KEY } from 'app/models/ApplicationDat
 import { CompanyAuthConfig } from 'app/models/CompanyAuthConfig';
 import { JwtEncryptionService } from 'app/modules/jwt-encryption-service/JwtEncryptionService';
 import { SESSION_NOT_FOUND_ERROR } from 'app/utils/CommonErrors';
+import { REVIEW_PENALTY_PAGE_URI, SELECT_THE_PENALTY_PAGE_URI } from 'app/utils/Paths';
 
 import { createSubstituteOf } from 'test/SubstituteFactory';
 
@@ -68,10 +73,10 @@ describe('Company Authentication Middleware', () => {
 
         const nextFunction = createSubstituteOf<NextFunction>();
         const response = createSubstituteOf<Response>();
-        const request = { session: undefined } as Request;
+        const req = { session: undefined } as Request;
 
         try{
-            await companyAuthMiddleware.handler(request, response, nextFunction);
+            await companyAuthMiddleware.handler(req, response, nextFunction);
         } catch (err) {
             expect(err).to.equal(SESSION_NOT_FOUND_ERROR);
         }
@@ -97,9 +102,9 @@ describe('Company Authentication Middleware', () => {
 
         const nextFunction = createSubstituteOf<NextFunction>();
         const response = createSubstituteOf<Response>();
-        const request = { session: undefined } as Request;
+        const req = { session: undefined } as Request;
 
-        await companyAuthMiddleware.handler(request, response, nextFunction);
+        await companyAuthMiddleware.handler(req, response, nextFunction);
         nextFunction.received(1);
         response.didNotReceive().redirect(Arg.any());
     });
@@ -124,9 +129,9 @@ describe('Company Authentication Middleware', () => {
 
         const nextFunction = createSubstituteOf<NextFunction>();
         const response = createSubstituteOf<Response>();
-        const request = getRequestSubstitute(appData, appeal.penaltyIdentifier.companyNumber);
+        const req = getRequestSubstitute(appData, appeal.penaltyIdentifier.companyNumber);
 
-        await companyAuthMiddleware.handler(request, response, nextFunction);
+        await companyAuthMiddleware.handler(req, response, nextFunction);
         nextFunction.received(1);
         response.didNotReceive().redirect(Arg.any());
 
@@ -138,7 +143,7 @@ describe('Company Authentication Middleware', () => {
 
         const nextFunction = createSubstituteOf<NextFunction>();
         const response = createSubstituteOf<Response>();
-        const request = getRequestSubstitute(appData, '');
+        const req = getRequestSubstitute(appData, '');
 
         const encryptionService = createSubstituteOf<JwtEncryptionService>(service => {
             service.encrypt(Arg.any()).resolves('MOCK');
@@ -156,10 +161,54 @@ describe('Company Authentication Middleware', () => {
 
         const startingWithRedirectUrl = (redirect : string) => redirect.startsWith(redirectUrl);
 
-        await companyAuthMiddleware.handler(request, response, nextFunction);
+        await companyAuthMiddleware.handler(req, response, nextFunction);
         nextFunction.didNotReceive();
         response.received(1).redirect(Arg.is(startingWithRedirectUrl));
 
+    });
+});
+
+describe('All pages after the Penalty Details page:', () => {
+
+    const pageList = [
+        ['Select Penalty', SELECT_THE_PENALTY_PAGE_URI],
+        ['Review Penalty', REVIEW_PENALTY_PAGE_URI]
+    ];
+
+    pageList.forEach((page) => {
+        it(`The ${page[0]} page should redirect unauthenticated user to the Company Auth service`, async () => {
+
+            const applicationData: Partial<ApplicationData> = {
+                appeal: {
+                    penaltyIdentifier: {
+                        companyNumber: 'NI999999',
+                        penaltyReference: 'PEN1A/NI000000',
+                        companyName: 'Test',
+                        penaltyList: {
+                            items: [
+                                {} as Penalty,
+                                {} as Penalty
+                            ]
+                        } as PenaltyList
+                    }
+                } as Appeal,
+                navigation: { permissions: [ page[1] ] }
+            };
+
+            /* TODO: Work out how to use this function call to override default Company Number
+            This will make the test data more clear, instead of changing the CN in the applicationData
+            above */
+            const app = createApp(applicationData);
+
+            await request(app)
+                .get(page[1])
+                .expect(res => {
+                    expect(res.status).to.be.equal(MOVED_TEMPORARILY);
+                    expect(res.text).to.contain('Found. Redirecting')
+                        .and.to.contain('oauth2/authorise')
+                        .and.to.contain('scope=https://api.companieshouse.gov.uk/company/NI999999');
+                });
+        });
     });
 });
